+++ apollo-composition/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## Unreleased
+
+- [#590](https://github.com/apollographql/federation-rs/pull/590) Fix
+  deserialization of `GraphQLError` nodes.
+
+## 0.1.4
+
+- Update to `apollo-federation` 2.0.0-preview.0
+
+## 0.1.3
+
+- [#586](https://github.com/apollographql/federation-rs/pull/586) Make
+  `SubgraphLocation.subgraph` an `Option`. For now, composition errors can have
+  no attributed subgraph.
+- [#583](https://github.com/apollographql/federation-rs/pull/583) Remove
+  connectors warning.
+
 ## 0.1.2
 
-* Updated dependencies
+- Updated dependencies