+++ apollo-federation-types/CHANGELOG.md
@@ -6,11 +6,18 @@ Not every version is listed here because versions before 0.14.0 did not have a c
 
 ### Breaking changes
 
+- `GraphQLError.nodes` is now an `Option<Vec<SubgraphASTNode>>`
 - All usages of `camino::Utf8PathBuf` have been replaced with `std::path::PathBuf`
 
 ### Features
 
 - A new `json_schema` feature derives the `schemars::JsonSchema` trait on `SupergraphConfig` and its sub-types.
+
+## 0.14.1 - 2024-09-19
+
+### Features
+
+- `impl FromIterator<(String, SubgraphConfig)> for SupergraphConfig`
 
 ## 0.14.0 - 2024-09-11